move EmailGenerator and EmailSender to logic.api class
This preserves the action -> logic.api dependency as specified in the design diagram
