+++ src/main/java/teammates/logic/core/InstructorsLogic.java
@@ -14,6 +14,8 @@ import teammates.common.util.EmailWrapper;
 import teammates.common.util.FieldValidator;
 import teammates.common.util.Logger;
 import teammates.common.util.StringHelper;
+import teammates.logic.api.EmailGenerator;
+import teammates.logic.api.EmailSender;
 import teammates.storage.api.InstructorsDb;
 
 /**
