+++ src/main/java/teammates/ui/automated/FeedbackSessionPublishedEmailWorkerAction.java
@@ -7,7 +7,7 @@ import teammates.common.exception.TeammatesException;
 import teammates.common.util.Assumption;
 import teammates.common.util.Const.ParamsNames;
 import teammates.common.util.EmailWrapper;
-import teammates.logic.core.EmailGenerator;
+import teammates.logic.api.EmailGenerator;
 import teammates.logic.core.FeedbackSessionsLogic;
 
 /**
