+++ src/main/java/teammates/ui/automated/PendingCommentClearedEmailWorkerAction.java
@@ -8,8 +8,8 @@ import teammates.common.exception.TeammatesException;
 import teammates.common.util.Assumption;
 import teammates.common.util.Const.ParamsNames;
 import teammates.common.util.EmailWrapper;
+import teammates.logic.api.EmailGenerator;
 import teammates.logic.core.CommentsLogic;
-import teammates.logic.core.EmailGenerator;
 import teammates.logic.core.FeedbackResponseCommentsLogic;
 
 /**
