+++ src/test/java/teammates/test/cases/logic/EmailGeneratorTest.java
@@ -18,8 +18,8 @@ import teammates.common.datatransfer.UserType;
 import teammates.common.util.Config;
 import teammates.common.util.EmailType;
 import teammates.common.util.EmailWrapper;
+import teammates.logic.api.EmailGenerator;
 import teammates.logic.core.CoursesLogic;
-import teammates.logic.core.EmailGenerator;
 import teammates.logic.core.FeedbackSessionsLogic;
 import teammates.logic.core.InstructorsLogic;
 import teammates.logic.core.StudentsLogic;
