+++ src/test/java/teammates/test/cases/logic/EmailSenderTest.java
@@ -11,7 +11,7 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import teammates.common.util.EmailWrapper;
-import teammates.logic.core.EmailSender;
+import teammates.logic.api.EmailSender;
 import teammates.logic.core.JavamailService;
 import teammates.logic.core.MailgunService;
 import teammates.logic.core.MailjetService;
